requirements.txt
discord.py==1.6.0